[Bronze IV] Title: SciComLove (2024), Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ C++17/백준/Bronze/31746. SciComLove （2024）/SciComLove （2024）.cc
@@ -7,7 +7,7 @@ int main() {
     cin.tie(0);
     cout.tie(0);
 
-    short n; cin >> n;
+    int n; cin >> n;
 
     if(n % 2 == 0) {
         cout << "SciComLove";
